add more media methods
