+++ src/main/java/com/jwplayer/jwplatform/client/AdvertisingClient.java
@@ -1,4 +1,4 @@
-package com.jwplayer.jwplatform;
+package com.jwplayer.jwplatform.client;
 
 public class AdvertisingClient{
